requirements.txt
flask
flask-cors
python-dotenv
google-genai
pymongo
werkzeug